Functionality 1.) User Authentication
Note
User authentication using x-auth

+++ models/todo.js
@@ -14,10 +14,10 @@ var getUUID = function() {
   };
   
   
-  var insertTodoData = function(uuid, todo) {
+  var insertTodoData = function(uuid, todo, uid) {
      var promise = new Promise(function(resolve, reject){
-        var sql = 'insert into toto (id, todo_text, comp_flag, cdate, udate) values (?, ?, 0, NOW(), NOW())';
-        mysql.dbCon.query(sql, [uuid, todo.text], (err, rows, fields) => {
+        var sql = 'insert into toto (id, uid, todo_text, comp_flag, cdate, udate) values (?, ?, ?, 0, NOW(), NOW())';
+        mysql.dbCon.query(sql, [uuid, uid, todo.text], (err, rows, fields) => {
             if (!err){
                 resolve({
                     data: uuid,
@@ -36,10 +36,10 @@ var getUUID = function() {
      return promise;
   };
 
-  var updateTodoData = function(textStr, id){
+  var updateTodoData = function(textStr, id, uid){
       var promise = new Promise((resolve, reject) => {
-        var sql = 'update toto set todo_text = ?, udate = NOW() where id = ?';
-        mysql.dbCon.query(sql, [textStr, id], (err, rows, fields) => {
+        var sql = 'update toto set todo_text = ?, udate = NOW() where id = ? and uid = ?';
+        mysql.dbCon.query(sql, [textStr, id, uid], (err, rows, fields) => {
             if(!err){
                 resolve({
                     id: id,
@@ -57,10 +57,10 @@ var getUUID = function() {
       return promise;
   };
 
-  var updateTodoCompletedData = function(id){
+  var updateTodoCompletedData = function(id, uid){
     var promise = new Promise((resolve, reject) => {
-        var sql = 'update toto set comp_flag = 1, comp_date = NOW() where id = ?';
-        mysql.dbCon.query(sql, [id], (err, rows, fields) => {
+        var sql = 'update toto set comp_flag = 1, comp_date = NOW() where id = ? and uid = ?';
+        mysql.dbCon.query(sql, [id, uid], (err, rows, fields) => {
             if(!err){
                 resolve({
                     id: id,
@@ -78,21 +78,21 @@ var getUUID = function() {
       return promise;
   };
 
-var insertTodo = (todo) => {
+var insertTodo = (todo, uid) => {
 
    return new Promise((resolve, reject) => {
         getUUID().then((data) => {
-            insertTodoData(data.uuid, todo).then((data1) => {
+            insertTodoData(data.uuid, todo, uid).then((data1) => {
                 resolve(data1);
             });
         });
     }); 
 };
 
-var updateTodo = (textStr, id) => {
+var updateTodo = (textStr, id, uid) => {
     return new Promise((resolve, reject) => {
-        updateTodoData(textStr, id).then((updateData) => {
-            getTodoById(id).then((data) => {
+        updateTodoData(textStr, id, uid).then((updateData) => {
+            getTodoById(id, uid).then((data) => {
                 resolve(data);
             });
         }).catch((err) => {
@@ -101,10 +101,10 @@ var updateTodo = (textStr, id) => {
     });
 };
 
-var updateTodoCompleted = (id) => {
+var updateTodoCompleted = (id, uid) => {
     return new Promise((resolve, reject) => {
-        updateTodoCompletedData(id).then((updateData) => {
-            getTodoById(id).then((data) => {
+        updateTodoCompletedData(id, uid).then((updateData) => {
+            getTodoById(id, uid).then((data) => {
                 resolve(data);
             });
         }).catch((err) => {
@@ -113,11 +113,11 @@ var updateTodoCompleted = (id) => {
     });
 };
 
-var getAllTodo = () => {
+var getAllTodo = (uid) => {
 
     return new Promise((resolve, reject) => {
-        var sql = 'SELECT * FROM toto where comp_flag = 0 Order by cdate';
-        mysql.dbCon.query(sql, (err, rows, fields) => {
+        var sql = 'SELECT * FROM toto where uid = ? and comp_flag = 0 Order by cdate';
+        mysql.dbCon.query(sql, [uid], (err, rows, fields) => {
 
             if (!err){
                 resolve({
@@ -137,11 +137,11 @@ var getAllTodo = () => {
     });
 };
 
-var getAllCompletedTodo = () => {
+var getAllCompletedTodo = (uid) => {
 
     return new Promise((resolve, reject) => {
-        var sql = 'SELECT * FROM toto where comp_flag = 1 Order by comp_date';
-        mysql.dbCon.query(sql, (err, rows, fields) => {
+        var sql = 'SELECT * FROM toto where uid = ? and comp_flag = 1 Order by comp_date';
+        mysql.dbCon.query(sql, [uid], (err, rows, fields) => {
 
             if (!err){
                 resolve({
@@ -161,12 +161,12 @@ var getAllCompletedTodo = () => {
     });
 };
 
-var getTodoById = (id) => {
+var getTodoById = (id, uid) => {
     
     return new Promise((resolve, reject) => {
         
-        var sql = 'SELECT * FROM toto where id = ?';
-        mysql.dbCon.query(sql, [id], (err, rows, fields) => {
+        var sql = 'SELECT * FROM toto where id = ? and uid = ?';
+        mysql.dbCon.query(sql, [id, uid], (err, rows, fields) => {
 
             if (!err){
                 resolve({
@@ -186,10 +186,10 @@ var getTodoById = (id) => {
     });
 };
 
-var deleteTodo = (id) => {
+var deleteTodo = (id, uid) => {
     return new Promise((resolve, reject) => {
-        var sql = 'delete from toto where id = ?';
-        mysql.dbCon.query(sql, [id], (err, rows, fields) => {
+        var sql = 'delete from toto where id = ? and uid = ?';
+        mysql.dbCon.query(sql, [id, uid], (err, rows, fields) => {
             if(!err){
                 resolve({
                     data: 'Todo deleted',
